+++ apps/bot/src/domains/crew/service.ts
@@ -1,4 +1,4 @@
-import { db, type CharacterInstance, type Player } from '@one-piece/db';
+import { db, type Player } from '@one-piece/db';
 
 import { NotFoundError, ValidationError } from '../../discord/errors.js';
 import { sanitizeName } from '../../shared/sanitize-name.js';
@@ -6,12 +6,12 @@ import * as characterRepository from '../character/repository.js';
 import type { CharacterRow } from '../character/types.js';
 import * as historyRepository from '../history/index.js';
 import * as playerRepository from '../player/repository.js';
-import * as shipRepository from '../ship/repository.js';
 
 import { MAX_CREW_NAME_LENGTH, MIN_CREW_NAME_LENGTH } from './constants.js';
 import * as crewRepository from './repository.js';
-import { getCrewCapacity } from './utils/get-crew-capacity.js';
 import { isInCrewFilter } from './utils/is-in-crew-filter.js';
+
+export { disembarkCharacter, embarkCharacter } from './services/crew-membership.js';
 
 export async function getCrewByPlayerId(playerId: number): Promise<Array<CharacterRow>> {
   const characters = await characterRepository.getCharactersByPlayerId(playerId);
@@ -22,34 +22,6 @@ export async function getCrewByPlayerId(playerId: number): Promise<Array<Charact
   }
 
   return crew;
-}
-
-export async function embarkCharacter(playerId: number, instanceId: number): Promise<void> {
-  await db.transaction(async (transaction) => {
-    const character = await crewRepository.findCharacterInstanceById(instanceId, transaction, { forUpdate: true });
-    assertCharacterBelongsToPlayer(character, playerId);
-    assertCharacterIsInReserve(character);
-
-    const ship = await shipRepository.findByPlayerIdOrThrow(playerId, transaction, { forUpdate: true });
-    const characters = await characterRepository.getCharactersByPlayerId(playerId, transaction);
-    const crewSize = characters.filter(isInCrewFilter).length;
-    if (crewSize >= getCrewCapacity(ship)) {
-      throw new ValidationError("Ton équipage est plein, débarque quelqu'un d'abord.");
-    }
-
-    await crewRepository.setCharacterCrewMembership(instanceId, new Date(), transaction);
-  });
-}
-
-export async function disembarkCharacter(playerId: number, instanceId: number): Promise<void> {
-  await db.transaction(async (transaction) => {
-    const character = await crewRepository.findCharacterInstanceById(instanceId, transaction, { forUpdate: true });
-    assertCharacterBelongsToPlayer(character, playerId);
-    assertCharacterIsInCrew(character);
-    assertCharacterIsNotCaptain(character);
-
-    await crewRepository.setCharacterCrewMembership(instanceId, null, transaction);
-  });
 }
 
 export async function replaceCaptainOfPlayer(playerId: number, instanceId: number): Promise<void> {
@@ -87,31 +59,4 @@ export async function renameCrew(playerId: number, rawName: string): Promise<Pla
   }
 
   return playerRepository.updateCrewName(playerId, sanitizedName);
-}
-
-function assertCharacterBelongsToPlayer(
-  character: CharacterInstance | undefined,
-  playerId: number,
-): asserts character is CharacterInstance {
-  if (character?.playerId === playerId) return;
-
-  throw new ValidationError("Ce personnage ne t'appartient pas.");
-}
-
-function assertCharacterIsInReserve(character: { joinedCrewAt: Date | null }): void {
-  if (character.joinedCrewAt === null) return;
-
-  throw new ValidationError('Ce personnage est déjà dans ton équipage.');
-}
-
-function assertCharacterIsInCrew(character: { joinedCrewAt: Date | null }): void {
-  if (character.joinedCrewAt !== null) return;
-
-  throw new ValidationError("Ce personnage n'est pas dans ton équipage.");
-}
-
-function assertCharacterIsNotCaptain(character: { isCaptain: boolean }): void {
-  if (!character.isCaptain) return;
-
-  throw new ValidationError("Réassigne le capitaine d'abord avec !changecaptain.");
 }
